+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:customizable_flashcard/flashcard_side_enum.dart';
 import 'package:flutter/material.dart';
 import 'package:customizable_flashcard/customizable_flashcard.dart';
 
@@ -5,22 +6,47 @@ void main() {
   runApp(const MainApp());
 }
 
-class MainApp extends StatelessWidget {
+class MainApp extends StatefulWidget {
   const MainApp({super.key});
 
+  @override
+  State<MainApp> createState() => _MainAppState();
+}
+
+class _MainAppState extends State<MainApp> {
+  bool isTapped = false;
+  FlashCardSide side = FlashCardSide.front;
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         body: Center(
-          child: FlashCard(
-              ontap: () {},
-              frontWidget: const Text("Front"),
-              backWidget: const Text("Back"),
-              frontColor: Colors.red,
-              backGradient: LinearGradient(
-                colors: [Colors.red, Colors.blue],
-              )),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              FlashCard(
+                  ontap: () {
+                    setState(() {
+                      isTapped = true;
+                    });
+                  },
+                  onFlip: (side) {
+                    setState(() {
+                      side = side;
+                    });
+                  },
+                  frontWidget: const Center(child: Text("Front")),
+                  backWidget: const Center(child: Text("Back")),
+                  frontGradient: const LinearGradient(
+                    colors: [Colors.red, Colors.blue],
+                  ),
+                  backGradient: const LinearGradient(
+                    colors: [Colors.red, Colors.blue],
+                  )),
+              Text("isPressed: $isTapped"),
+              Text("current side: $side"),
+            ],
+          ),
         ),
       ),
     );